minor fixes involving is_byte()
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@2262 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/utils.c
@@ -67,9 +67,9 @@ int num_of_digits(int n)
     return i;
 }
 
-bool is_byte(int c)
+bool is_byte(unsigned int c)
 {
-    return ((unsigned int)c == (unsigned char)c);
+    return (c == (unsigned char)c);
 }
 
 /* Read a ssize_t from str, and store it in *val (if val is not NULL).
